Upgrade Jekyll and add webrick
Get Jekyll Running locally on Ruby 3.1.0

In this branch Jekyll is upgraded to 4.2. WEBrick was removed
in Ruby 3 I think so I added it to the gemfile to have jekyll
start locally.

There were a couple of YAML exceptions that prompted me to change
the description key in index.html, and also in the _config.yml file.
I'm not sure which the description are different in index.html and
_config.yml files.